tests/wpt/tests/html/canvas/offscreen/canvas-host/2d.canvas.host.size.invalid.attributes.idl.html

<!DOCTYPE html>
<!-- DO NOT EDIT! This test has been generated by /html/canvas/tools/gentest.py. -->
<title>OffscreenCanvas test: 2d.canvas.host.size.invalid.attributes.idl</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/html/canvas/resources/canvas-tests.js"></script>

<h1>2d.canvas.host.size.invalid.attributes.idl</h1>
<p class="desc">Getting/setting width/height IDL attributes</p>


<script>
var t = async_test("Getting/setting width/height IDL attributes");
var t_pass = t.done.bind(t);
var t_fail = t.step_func(function(reason) {
    throw reason;
});
t.step(function() {

  var canvas = new OffscreenCanvas(100, 50);
  var ctx = canvas.getContext('2d');

  assert_throws_js(TypeError, function() { canvas.width = 200 - Math.pow(2, 32); });
  assert_throws_js(TypeError, function() { canvas.height = 200 - Math.pow(2, 32); });
  assert_throws_js(TypeError, function() { canvas.width = '400x'; });
  assert_throws_js(TypeError, function() { canvas.height = 'foo'; });
  t.done();

});
</script>
